Limited separation of transaction orchestration and participation.
