+++ src/Orleans.Transactions/State/ReaderWriterLock.cs
@@ -3,15 +3,22 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Orleans.Configuration;
 using Orleans.Runtime;
-using Orleans.Transactions;
 using Orleans.Transactions.Abstractions;
 
-namespace Orleans.Transactions
+namespace Orleans.Transactions.State
 {
-    public partial class TransactionalState<TState> : ITransactionalState<TState>, ITransactionParticipant, ILifecycleParticipant<IGrainLifecycle>
+    internal class ReadWriteLock<TState>
        where TState : class, new()
     {
+        private readonly TransactionalStateOptions options;
+        private readonly TransactionQueue<TState> queue;
+        private BatchWorker lockWorker;
+        private BatchWorker storageWorker;
+        private readonly ILogger logger;
+
         // the linked list of lock groups
         // the head is the group that is currently holding the lock
         private LockGroup currentGroup = null;
@@ -30,9 +37,218 @@ namespace Orleans.Transactions
             public DateTime? Deadline;
         }
 
-        // check for transactions in the lock stage that can exit it, 
-        // and for transactions in the wait stage that can enter the lock stage,
-        // and for expired group lock
+        public ReadWriteLock(
+            IOptions<TransactionalStateOptions> options,
+            TransactionQueue<TState> queue,
+            BatchWorker storageWorker,
+            ILogger logger)
+        {
+            this.options = options.Value;
+            this.queue = queue;
+            this.storageWorker = storageWorker;
+            this.logger = logger;
+            this.lockWorker = new BatchWorkerFromDelegate(LockWork);
+        }
+
+        public Task<TResult> EnterLock<TResult>(Guid transactionId, DateTime priority,
+                                   AccessCounter counter, bool isRead, Task<TResult> task)
+        {
+            bool rollbacksOccurred = false;
+
+            // search active transactions
+            if (Find(transactionId, isRead, out var group, out var record))
+            {
+                // check if we lost some reads or writes already
+                if (counter.Reads > record.NumberReads || counter.Writes > record.NumberWrites)
+                {
+                    throw new OrleansBrokenTransactionLockException(transactionId.ToString(), "when re-entering lock");
+                }
+
+                // check if the operation conflicts with other transactions in the group
+                if (HasConflict(isRead, priority, transactionId, group, out var resolvable))
+                {
+                    if (!resolvable)
+                    {
+                        throw new OrleansTransactionLockUpgradeException(transactionId.ToString());
+                    }
+                    else
+                    {
+                        // rollback all conflicts
+                        var conflicts = Conflicts(transactionId, group).ToList();
+
+                        if (conflicts.Count > 0)
+                        {
+                            foreach (var r in conflicts)
+                            {
+                                Rollback(r, true);
+                                rollbacksOccurred = true;
+                            }
+                        }
+                    }
+                }
+            }
+            else
+            {
+                // check if we were supposed to already hold this lock
+                if (counter.Reads + counter.Writes > 0)
+                {
+                    throw new OrleansBrokenTransactionLockException(transactionId.ToString(), "when trying to re-enter lock");
+                }
+
+                // update the lock deadline
+                if (group == currentGroup)
+                {
+                    group.Deadline = DateTime.UtcNow + this.options.LockTimeout;
+                }
+
+                // create a new record for this transaction
+                record = new TransactionRecord<TState>()
+                {
+                    TransactionId = transactionId,
+                    Priority = priority,
+                    Deadline = DateTime.UtcNow + this.options.LockAcquireTimeout
+                };
+
+                group.Add(transactionId, record);
+                group.FillCount++;
+
+                if (logger.IsEnabled(LogLevel.Trace))
+                {
+                    if (group == currentGroup)
+                        logger.Trace($"enter-lock {transactionId} fc={group.FillCount}");
+                    else
+                        logger.Trace($"enter-lock-queue {transactionId} fc={group.FillCount}");
+                }
+            }
+
+            if (group != currentGroup)
+            {
+                // task will be executed once its group acquires the lock
+
+                if (group.Tasks == null)
+                    group.Tasks = new List<Task>();
+
+                group.Tasks.Add(task);
+            }
+            else
+            {
+                // execute task right now
+                task.RunSynchronously();
+
+                // look at exception to avoid UnobservedException
+                var ignore = task.Exception;
+            }
+
+            if (isRead)
+            {
+                record.AddRead();
+            }
+            else
+            {
+                record.AddWrite();
+            }
+
+            if (rollbacksOccurred)
+            {
+                lockWorker.Notify();
+            }
+            else if (group.Deadline.HasValue)
+            {
+                lockWorker.Notify(group.Deadline.Value);
+            }
+
+            return task;
+        }
+
+        public bool ValidateLock(Guid transactionId, AccessCounter accessCount, out TransactionalStatus status, out TransactionRecord<TState> record)
+        {
+            if (currentGroup == null || !currentGroup.TryGetValue(transactionId, out record))
+            {
+                record = new TransactionRecord<TState>()
+                {
+                    TransactionId = transactionId
+                };
+                status = TransactionalStatus.BrokenLock;
+                return false;
+            }
+            else if (record.NumberReads != accessCount.Reads
+                   || record.NumberWrites != accessCount.Writes)
+            {
+                Rollback(transactionId, true);
+
+                status = TransactionalStatus.LockValidationFailed;
+                return false;
+            }
+            else
+            {
+                status = TransactionalStatus.Ok;
+                return true;
+            }
+        }
+
+        public void Notify()
+        {
+            this.lockWorker.Notify();
+        }
+
+        public bool TryGetRecord(Guid transactionId, out TransactionRecord<TState> record)
+        {
+            return this.currentGroup.TryGetValue(transactionId, out record);
+        }
+
+        public void AbortExecutingTransactions()
+        {
+            if (currentGroup != null)
+            {
+                foreach (var kvp in currentGroup)
+                {
+                    if (logger.IsEnabled(LogLevel.Trace))
+                        logger.Trace($"break-lock for transaction {kvp.Key}");
+
+                    this.queue.NotifyOfAbort(kvp.Value, TransactionalStatus.BrokenLock);
+                }
+
+                currentGroup.Clear();
+            }
+        }
+
+        public void AbortQueuedTransactions()
+        {
+            var pos = currentGroup?.Next;
+            while (pos != null)
+            {
+                if (pos.Tasks != null)
+                {
+                    foreach (var t in pos.Tasks)
+                    {
+                        // running the task will abort the transaction because it is not in currentGroup
+                        t.RunSynchronously();
+                        // look at exception to avoid UnobservedException
+                        var ignore = t.Exception;
+                    }
+                }
+                pos = pos.Next;
+            }
+        }
+
+        public void Rollback(Guid guid, bool notify)
+        {
+            // no-op if the transaction never happened or already rolled back
+            if (currentGroup == null || !currentGroup.TryGetValue(guid, out var record))
+            {
+                return;
+            }
+
+            // notify remote listeners
+            if (notify)
+            {
+                this.queue.NotifyOfAbort(record, TransactionalStatus.BrokenLock);
+            }
+
+            // remove record for this transaction
+            currentGroup.Remove(guid);
+        }
+
         private Task LockWork()
         {
             logger.Trace("/LockWork");
@@ -46,13 +262,13 @@ namespace Orleans.Transactions
                     {
                         if (single != null)
                         {
-                            EnqueueCommit(single);
+                            this.queue.EnqueueCommit(single);
                         }
                         else if (multiple != null)
                         {
                             foreach (var r in multiple)
                             {
-                                EnqueueCommit(r);
+                                this.queue.EnqueueCommit(r);
                             }
                         }
 
@@ -65,7 +281,7 @@ namespace Orleans.Transactions
                         // the lock group has timed out.
                         var txlist = string.Join(",", currentGroup.Keys.Select(g => g.ToString()));
                         logger.Warn(555, $"break-lock timeout for {currentGroup.Count} transactions {txlist}");
-                        AbortExecutingTransactions("after lock timeout");
+                        AbortExecutingTransactions();
                         lockWorker.Notify();
                     }
 
@@ -83,7 +299,7 @@ namespace Orleans.Transactions
 
                     if (currentGroup != null)
                     {
-                        currentGroup.Deadline = DateTime.UtcNow + LockTimeout;
+                        currentGroup.Deadline = DateTime.UtcNow + this.options.LockTimeout;
 
                         // discard expired waiters that have no chance to succeed
                         // because they have been waiting for the lock for a longer timespan than the 
@@ -138,120 +354,7 @@ namespace Orleans.Transactions
 
             return Task.CompletedTask;
         }
-
-
-
-        // blocks until the given operation for this transaction can be executed.
-        private Task<TResult> EnterLock<TResult>(Guid transactionId, DateTime priority,
-                                   AccessCounter counter, bool isRead, Task<TResult> task)
-        {
-            bool rollbacksOccurred = false;
-
-            // search active transactions
-            if (Find(transactionId, isRead, out var group, out var record))
-            {
-                // check if we lost some reads or writes already
-                if (counter.Reads > record.NumberReads || counter.Writes > record.NumberWrites)
-                {
-                    throw new OrleansBrokenTransactionLockException(transactionId.ToString(), "when re-entering lock");
-                }
-
-                // check if the operation conflicts with other transactions in the group
-                if (HasConflict(isRead, priority, transactionId, group, out var resolvable))
-                {
-                    if (!resolvable)
-                    {
-                        throw new OrleansTransactionLockUpgradeException(transactionId.ToString());
-                    }
-                    else
-                    {
-                        // rollback all conflicts
-                        var conflicts = Conflicts(transactionId, group).ToList();
-
-                        if (conflicts.Count > 0)
-                        {
-                            foreach (var r in conflicts)
-                            {
-                                Rollback(r, "wait-die on conflict", true);
-                                rollbacksOccurred = true;
-                            }
-                        }
-                    }
-                }
-            }
-            else
-            {
-                // check if we were supposed to already hold this lock
-                if (counter.Reads + counter.Writes > 0)
-                {
-                    throw new OrleansBrokenTransactionLockException(transactionId.ToString(), "when trying to re-enter lock");
-                }
-
-                // update the lock deadline
-                if (group == currentGroup)
-                {
-                    group.Deadline = DateTime.UtcNow + LockTimeout;
-                }
-
-                // create a new record for this transaction
-                record = new TransactionRecord<TState>()
-                {
-                    TransactionId = transactionId,
-                    Priority = priority,
-                    Deadline = DateTime.UtcNow + LockAcquireTimeout
-                };
-
-                group.Add(transactionId, record);
-                group.FillCount++;
-
-                if (logger.IsEnabled(LogLevel.Trace))
-                {
-                    if (group == currentGroup)
-                        logger.Trace($"enter-lock {transactionId} fc={group.FillCount}");
-                    else
-                        logger.Trace($"enter-lock-queue {transactionId} fc={group.FillCount}");
-                }
-            }
-
-            if (group != currentGroup)
-            {
-                // task will be executed once its group acquires the lock
-
-                if (group.Tasks == null)
-                    group.Tasks = new List<Task>();
-
-                group.Tasks.Add(task);
-            }
-            else
-            {
-                // execute task right now
-                task.RunSynchronously();
-
-                // look at exception to avoid UnobservedException
-                var ignore = task.Exception;
-            }
-
-            if (isRead)
-            {
-                record.AddRead();
-            }
-            else
-            {
-                record.AddWrite();
-            }
-
-            if (rollbacksOccurred)
-            {
-                lockWorker.Notify();
-            }
-            else if (group.Deadline.HasValue)
-            {
-                lockWorker.Notify(group.Deadline.Value);
-            }
-
-            return task;
-        }
-
+       
         private bool Find(Guid guid, bool isRead, out LockGroup group, out TransactionRecord<TState> record)
         {
             if (currentGroup == null)
@@ -336,32 +439,6 @@ namespace Orleans.Transactions
                 {
                     yield return kvp.Key;
                 }
-            }
-        }
-
-        private bool ValidateLock(Guid transactionId, AccessCounter accessCount, out TransactionalStatus status, out TransactionRecord<TState> record)
-        {
-            if (currentGroup == null || !currentGroup.TryGetValue(transactionId, out record))
-            {
-                record = new TransactionRecord<TState>()
-                {
-                    TransactionId = transactionId
-                };
-                status = TransactionalStatus.BrokenLock;
-                return false;
-            }
-            else if (record.NumberReads != accessCount.Reads
-                   || record.NumberWrites != accessCount.Writes)
-            {
-                Rollback(transactionId, "access count mismatch on prepare", true);
-
-                status = TransactionalStatus.LockValidationFailed;
-                return false;
-            }
-            else
-            {
-                status = TransactionalStatus.Ok;
-                return true;
             }
         }
 
@@ -453,68 +530,5 @@ namespace Orleans.Transactions
         {
             return a.Timestamp.CompareTo(b.Timestamp);
         }
-
-        // aborts all executing transactions
-        private void AbortExecutingTransactions(string indication)
-        {
-            if (currentGroup != null)
-            {
-                foreach (var kvp in currentGroup)
-                {
-                    if (logger.IsEnabled(LogLevel.Trace))
-                        logger.Trace($"break-lock {indication} for transaction {kvp.Key}");
-
-                    NotifyOfAbort(kvp.Value, TransactionalStatus.BrokenLock);
-                }
-
-                currentGroup.Clear();
-            }
-        }
-
-        // aborts all queued transactions that are not in the lock yet
-        private void AbortQueuedTransactions()
-        {
-            var pos = currentGroup?.Next;
-            while (pos != null)
-            {
-                if (pos.Tasks != null)
-                {
-                    foreach (var t in pos.Tasks)
-                    {
-                        // running the task will abort the transaction because it is not in currentGroup
-                        t.RunSynchronously();
-                        // look at exception to avoid UnobservedException
-                        var ignore = t.Exception;
-                    }
-                }
-                pos = pos.Next;
-            }
-        }
-
-
-        // aborts transaction, if still active
-        private void Rollback(Guid guid, string indication, bool notify)
-        {
-            // no-op if the transaction never happened or already rolled back
-            if (currentGroup == null || !currentGroup.TryGetValue(guid, out var record))
-            {
-                return;
-            }
-
-            if (logger.IsEnabled(LogLevel.Trace))
-                logger.Trace($"break-lock {indication} for transaction {guid}");
-
-            // notify remote listeners
-            if (notify)
-            {
-                NotifyOfAbort(record, TransactionalStatus.BrokenLock);
-            }
-
-            // remove record for this transaction
-            currentGroup.Remove(guid);
-        }
-
     }
-
-
 }
